Перенёс все конверторы поближе к моделькам gorm-models. У меня не все конвертеры, если что нужно добавить добавлю XD
